requirements.txt
transformers
torch
pandas
nltk
scikit-learn
huggingface_hub
bitsandbytes
accelerate
google-generativeai
python-dotenv 